new hopefully fixed CustomUserEvents
